+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/pipe/task/subtask/connector/PipeConnectorSubtask.java
@@ -19,7 +19,6 @@
 
 package org.apache.iotdb.db.pipe.task.subtask.connector;
 
-import org.apache.iotdb.commons.exception.pipe.PipeRuntimeConnectorCriticalException;
 import org.apache.iotdb.commons.exception.pipe.PipeRuntimeConnectorRetryTimesConfigurableException;
 import org.apache.iotdb.commons.exception.pipe.PipeRuntimeException;
 import org.apache.iotdb.commons.pipe.config.PipeConfig;
@@ -158,76 +157,6 @@ public class PipeConnectorSubtask extends PipeTransferSubtask {
 
     event.onTransferred();
     PipeConnectorMetrics.getInstance().markPipeHeartbeatEvent(taskID);
-  }
-
-  /** @return true if the pipe task should be stopped, false otherwise */
-  private boolean onPipeConnectionException(Throwable throwable) {
-    LOGGER.warn(
-        "PipeConnectionException occurred, {} retries to handshake with the target system.",
-        outputPipeConnector.getClass().getName(),
-        throwable);
-
-    int retry = 0;
-    while (retry < MAX_RETRY_TIMES) {
-      try {
-        outputPipeConnector.handshake();
-        LOGGER.info(
-            "{} handshakes with the target system successfully.",
-            outputPipeConnector.getClass().getName());
-        break;
-      } catch (Exception e) {
-        retry++;
-        LOGGER.warn(
-            "{} failed to handshake with the target system for {} times, "
-                + "will retry at most {} times.",
-            outputPipeConnector.getClass().getName(),
-            retry,
-            MAX_RETRY_TIMES,
-            e);
-        try {
-          Thread.sleep(retry * PipeConfig.getInstance().getPipeConnectorRetryIntervalMs());
-        } catch (InterruptedException interruptedException) {
-          LOGGER.info(
-              "Interrupted while sleeping, will retry to handshake with the target system.",
-              interruptedException);
-          Thread.currentThread().interrupt();
-        }
-      }
-    }
-
-    // Stop current pipe task directly if failed to reconnect to
-    // the target system after MAX_RETRY_TIMES times
-    if (retry == MAX_RETRY_TIMES && lastEvent instanceof EnrichedEvent) {
-      report(
-          (EnrichedEvent) lastEvent,
-          new PipeRuntimeConnectorCriticalException(
-              throwable.getMessage()
-                  + ", root cause: "
-                  + ErrorHandlingUtils.getRootCause(throwable).getMessage()));
-      LOGGER.warn(
-          "{} failed to handshake with the target system after {} times, "
-              + "stopping current subtask {} (creation time: {}, simple class: {}). "
-              + "Status shown when query the pipe will be 'STOPPED'. "
-              + "Please restart the task by executing 'START PIPE' manually if needed.",
-          outputPipeConnector.getClass().getName(),
-          MAX_RETRY_TIMES,
-          taskID,
-          creationTime,
-          this.getClass().getSimpleName(),
-          throwable);
-
-      // Although the pipe task will be stopped, we still don't release the last event here
-      // Because we need to keep it for the next retry. If user wants to restart the task,
-      // the last event will be processed again. The last event will be released when the task
-      // is dropped or the process is running normally.
-
-      // Stop current pipe task if failed to reconnect to the target system after MAX_RETRY_TIMES
-      return true;
-    }
-
-    // For non enriched event, forever retry.
-    // For enriched event, retry if connection is set up successfully.
-    return false;
   }
 
   @Override
