+++ iotdb-core/node-commons/src/main/java/org/apache/iotdb/commons/pipe/task/subtask/PipeReportableSubtask.java
@@ -58,7 +58,6 @@ public abstract class PipeReportableSubtask extends PipeSubtask {
   }
 
   private void onEnrichedEventFailure(Throwable throwable) {
-
     int maxRetryTimes =
         throwable instanceof PipeRuntimeConnectorRetryTimesConfigurableException
             ? ((PipeRuntimeConnectorRetryTimesConfigurableException) throwable).getRetryTimes()
